Introduce the symbol clustering algorithm
This change improves the resolution of "jump to definition" and "get
usages"-like queries by enhancing accuracy from linkage information.
Symbols which are static-linked together to a query are shown in bold
green, symbol that aren't will be grey. Therefore, users can better
choose which "definition" a call site is referring to better, for large
projects.

Signed-off-by: Szalay Richárd <szalayrichard@inf.elte.hu>